and lo, he used two pointers, and it was good
FIRST the message came down from the mountain,
"use two pointers"
But the implementation was not followed, and there was but one pointer,
and the null pointer exception ensued.
SECOND the message came down again,
"use two pointers"
But the implementation was not followed, and there was but one pointer,
and the null pointer exception ensued.
AND SO the sage descended from the mountain,
and used two pointers
and it was good.